Fix errors
Fix error on overriding types
Fix error on empty where syntax

+++ src/classes/resolvers/DeleteResolver.js
@@ -1,11 +1,16 @@
 /* eslint-disable no-return-await */
 /* eslint-disable no-debugger */
 /* eslint-disable no-restricted-syntax */
-const Resolver = require('./Resolver');
+const Resolver = require("./Resolver");
+const _ = require("lodash");
 
 class DeleteResolver extends Resolver {
   async _resolve(parent, args, context, info) {
     const { where } = args;
+
+    if (_.isEmpty(where) === true) {
+      throw new Error("Invalid where syntax");
+    }
 
     const result = await this.model.destroy({ where });
     return result;
